ENH: added support for connected dropdowns related to folder suffixes.
Added support to allow for a two step dropdown in the cases selection.
A first dropdown will display values, and a second dropdown is populated
based on the selection of the first dropdown.  The second dropdown
is populated with suffixes specific to an input folder.  Also fixed
a bug as part of this where a suffix variable if optional was not
able to pass its value as the empty string, can now do so.